requirements.txt
flask
flask_restful 
flask_sqlalchemy 
flask_migrate
marshmallow-sqlalchemy
pymsql 
panda
numpy